Eliminitated the different node types
