+++ src/core/action.ts
@@ -18,9 +18,9 @@ function runRawAction(actioncall: IRawActionCall): any {
     return actioncall.object[actioncall.name].apply(actioncall.object, actioncall.args)
 }
 
-function collectMiddlewareHandlers(node: ComplexNode): IMiddleWareHandler[] {
+function collectMiddlewareHandlers(node: AbstractNode): IMiddleWareHandler[] {
     let handlers = node.middlewares.slice()
-    let n: ComplexNode = node
+    let n: AbstractNode = node
     // Find all middlewares. Optimization: cache this?
     while (n.parent) {
         n = n.parent
@@ -29,7 +29,7 @@ function collectMiddlewareHandlers(node: ComplexNode): IMiddleWareHandler[] {
     return handlers
 }
 
-function runMiddleWares(node: ComplexNode, baseCall: IRawActionCall): any {
+function runMiddleWares(node: AbstractNode, baseCall: IRawActionCall): any {
     const handlers = collectMiddlewareHandlers(node)
     // Short circuit
     if (!handlers.length)
@@ -76,7 +76,7 @@ export function createActionInvoker(name: string, fn: Function) {
     return createNamedFunction(name, actionInvoker)
 }
 
-function serializeArgument(node: ComplexNode, actionName: string, index: number, arg: any): any {
+function serializeArgument(node: AbstractNode, actionName: string, index: number, arg: any): any {
     if (isPrimitive(arg))
         return arg
     if (isComplexValue(arg)) {
@@ -103,7 +103,7 @@ function serializeArgument(node: ComplexNode, actionName: string, index: number,
     }
 }
 
-function deserializeArgument(adm: ComplexNode, value: any): any {
+function deserializeArgument(adm: AbstractNode, value: any): any {
     if (typeof value === "object") {
         const keys = Object.keys(value)
         if (keys.length === 1 && keys[0] === "$ref")
@@ -146,6 +146,6 @@ export function onAction(target: IComplexValue, listener: (call: ISerializedActi
     })
 }
 
-import { ComplexNode, getComplexNode,  IComplexValue, isComplexValue, } from "./nodes/complex-node"
+import { AbstractNode, getComplexNode,  IComplexValue, isComplexValue, } from "./nodes/abstract-node"
 import { resolve, tryResolve, addMiddleware,  } from "./mst-operations"
 import { fail, isPlainObject, isPrimitive, argsToArray, createNamedFunction, IDisposer } from "../utils"
