+++ src/core/index.ts
@@ -1,6 +1,4 @@
 export * from "./nodes/abstract-node"
-export * from "./nodes/complex-node"
-export * from "./nodes/immutable-node"
 export * from "./action"
 export * from "./json-patch"
 export * from "./mst-operations"
