+++ src/types/utility-types/literal.ts
@@ -11,11 +11,6 @@ export class Literal<T> extends Type<T, T> {
         this.value = value
     }
 
-    instantiate(parent: ComplexNode | null, subpath: string, environment: any, snapshot: T): AbstractNode {
-        typecheck(this, snapshot)
-        return new ImmutableNode(this, parent, subpath, this.value)
-    }
-
     describe() {
         return JSON.stringify(this.value)
     }
@@ -37,4 +32,4 @@ export function literal<S>(value: S): ISimpleType<S> {
     return new Literal<S>(value)
 }
 
-import { ImmutableNode, AbstractNode, ComplexNode } from '../../core';
+import { AbstractNode } from '../../core';
