+++ src/types/utility-types/reference.ts
@@ -1,7 +1,6 @@
 import { ISimpleType, TypeFlags, Type, IType } from "../type"
 import { IContext, IValidationResult, typeCheckSuccess, typeCheckFailure, typecheck } from "../type-checker"
 import { isPrimitive, fail } from "../../utils"
-import { ImmutableNode, ComplexNode, AbstractNode } from "../../core"
 
 
 export interface IReference {
@@ -18,11 +17,6 @@ export class ReferenceType<T> extends Type<ReferenceSnapshot, T> {
         private readonly basePath: string
     ) {
         super(`reference(${targetType.name})`)
-    }
-
-    instantiate(parent: ComplexNode, subpath: string, environment: any, snapshot: ReferenceSnapshot): AbstractNode {
-        typecheck(this.targetType, snapshot)
-        return new ImmutableNode(this, parent, subpath, snapshot)
     }
 
     describe() {
