+++ test/object.ts
@@ -43,8 +43,10 @@ const createTestFactories = () => {
 
 // === FACTORY TESTS ===
 test.only("it should create a factory", (t) => {
+    debugger;
     const {Factory} = createTestFactories()
-    const snapshot = getSnapshot(Factory.create())
+    const instance = Factory.create()
+    const snapshot = getSnapshot(instance)
     t.deepEqual(snapshot, {to: 'world'})
     t.deepEqual((Factory.create() as any).toJSON(), {to: 'world'}) // toJSON is there as shortcut for getSnapshot(), primarily for debugging convenience
     t.deepEqual(Factory.create().toString(), "AnonymousModel@<root>")
